Created priority inversion scenario
